requirements.txt
# PyTorch (CPU version)
torch==2.1.1 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.1 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.1.1 --index-url https://download.pytorch.org/whl/cpu

# YOLOv8
ultralytics==8.2.14

# Flask (for deployment API)
flask==2.3.3

# Utilities
opencv-python==4.9.0.80
matplotlib==3.8.4
numpy==1.26.4
tqdm==4.66.4
Pillow==10.3.0
